re-name api-client-extensions to node-api-extensions
* move the node API factory to that crate
* move the untrusted peer fetch back to the sidechain crate
* remove the new 'parentchain node api' crate
