+++ Cargo.toml
@@ -9,13 +9,12 @@ members = [
     "core/parentchain/block-importer",
     "core/parentchain/indirect-calls-executor",
     "core/parentchain/light-client",
-    "core/parentchain/node-api",
     "core/parentchain/parentchain-crate",
     "core/rest-client",
     "core/rpc-client",
     "core/rpc-server",
     "core/tls-websocket-server",
-    "core-primitives/api-client-extensions",
+    "core-primitives/node-api-extensions",
     "core-primitives/component-container",
     "core-primitives/enclave-api",
     "core-primitives/enclave-api/ffi",
