--- core/parentchain/node-api/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "itc-parentchain-node-api"
-version = "0.8.0"
-authors = ["Integritee AG <hello@integritee.network>"]
-edition = "2018"
-
-[dependencies]
-# crates.io
-thiserror = "1.0"
-# substrate
-sp-core = { version = "4.1.0-dev", git = "https://github.com/paritytech/substrate.git", branch = "master" }
-# scs / integritee
-substrate-api-client = { git = "https://github.com/scs/substrate-api-client", branch = "master" }
-# local
-itc-rpc-client = { path = "../../rpc-client" }
-itp-api-client-extensions = { path = "../../../core-primitives/api-client-extensions" }
-itp-types = { path = "../../../core-primitives/types" }
-
-[dev-dependencies]
-# crates.io
-mockall = "0.10.1"
-
-[features]
-mocks = []
