--- core/parentchain/node-api/src/error.rs
@@ -1,34 +1,0 @@
-/*
-	Copyright 2021 Integritee AG and Supercomputing Systems AG
-
-	Licensed under the Apache License, Version 2.0 (the "License");
-	you may not use this file except in compliance with the License.
-	You may obtain a copy of the License at
-
-		http://www.apache.org/licenses/LICENSE-2.0
-
-	Unless required by applicable law or agreed to in writing, software
-	distributed under the License is distributed on an "AS IS" BASIS,
-	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-	See the License for the specific language governing permissions and
-	limitations under the License.
-
-*/
-
-//! Node API error.
-
-use itp_types::ShardIdentifier;
-
-pub type Result<T> = core::result::Result<T, Error>;
-
-#[derive(Debug, thiserror::Error)]
-pub enum Error {
-	#[error("Node RPC client error: {0}")]
-	RpcClient(#[from] itc_rpc_client::error::Error),
-	#[error("Failed to create a node API: {0}")]
-	FailedToCreateNodeApi(#[from] substrate_api_client::ApiClientError),
-	#[error("Could not find any peers on-chain for shard: {0:?}")]
-	NoPeerFoundForShard(ShardIdentifier),
-	#[error(transparent)]
-	Other(#[from] Box<dyn std::error::Error + Sync + Send + 'static>),
-}
