+++ service/src/tests/parentchain_block_syncer_test.rs
@@ -22,7 +22,7 @@ use crate::{
 	},
 	ParentchainBlockSyncer,
 };
-use itp_api_client_extensions::ChainApi;
+use itp_node_api_extensions::ChainApi;
 use std::sync::Arc;
 
 #[test]
