+++ sidechain/peer-fetch/src/block_fetch_client.rs
@@ -15,9 +15,8 @@
 
 */
 
-use crate::{error::Result, FetchBlocksFromPeer};
+use crate::{error::Result, untrusted_peer_fetch::FetchUntrustedPeers, FetchBlocksFromPeer};
 use async_trait::async_trait;
-use itc_parentchain_node_api::untrusted_peer_fetch::FetchUntrustedPeers;
 use its_primitives::{
 	constants::RPC_METHOD_NAME_FETCH_BLOCKS_FROM_PEER,
 	traits::SignedBlock as SignedBlockTrait,
@@ -82,8 +81,10 @@ where
 mod tests {
 
 	use super::*;
-	use crate::block_fetch_server::BlockFetchServerModuleBuilder;
-	use itc_parentchain_node_api::mocks::untrusted_peer_fetch_mock::UntrustedPeerFetcherMock;
+	use crate::{
+		block_fetch_server::BlockFetchServerModuleBuilder,
+		mocks::untrusted_peer_fetch_mock::UntrustedPeerFetcherMock,
+	};
 	use its_primitives::types::SignedBlock;
 	use its_storage::fetch_blocks_mock::FetchBlocksMock;
 	use its_test::sidechain_block_builder::SidechainBlockBuilder;
