+++ sidechain/peer-fetch/src/error.rs
@@ -23,14 +23,16 @@ pub type Result<T> = core::result::Result<T, Error>;
 pub enum Error {
 	#[error("RPC client error: {0}")]
 	RpcClient(#[from] itc_rpc_client::error::Error),
-	#[error("Node API error: {0}")]
-	NodeApi(#[from] itc_parentchain_node_api::error::Error),
-	#[error("Node API extension error: {0}")]
-	NodeApiExtensions(#[from] itp_api_client_extensions::ApiClientError),
+	#[error("Node API extensions error: {0}")]
+	NodeApiExtensions(#[from] itp_node_api_extensions::ApiClientError),
+	#[error("Node API factory error: {0}")]
+	NodeApiFactory(#[from] itp_node_api_extensions::node_api_factory::NodeApiFactoryError),
 	#[error("Serialization error: {0}")]
 	Serialization(#[from] serde_json::Error),
 	#[error("JSON RPC error: {0}")]
 	JsonRpc(#[from] jsonrpsee::types::Error),
+	#[error("Could not find any peers on-chain for shard: {0:?}")]
+	NoPeerFoundForShard(its_primitives::types::ShardIdentifier),
 	#[error(transparent)]
 	Other(#[from] Box<dyn std::error::Error + Sync + Send + 'static>),
 }
