+++ sidechain/peer-fetch/src/untrusted_peer_fetch.rs
@@ -15,13 +15,10 @@
 
 */
 
-use crate::{
-	error::{Error, Result},
-	node_api_factory::CreateNodeApi,
-};
+use crate::error::{Error, Result};
 use itc_rpc_client::direct_client::{DirectApi, DirectClient as DirectWorkerApi};
-use itp_api_client_extensions::PalletTeerexApi;
-use itp_types::ShardIdentifier;
+use itp_node_api_extensions::{node_api_factory::CreateNodeApi, PalletTeerexApi};
+use its_primitives::types::ShardIdentifier;
 use std::sync::Arc;
 
 /// Trait to fetch untrusted peer servers.
